deconstruct commons/models.py into separate modules
