+++ src/common/gesture_action_mapper.py
@@ -1,4 +1,4 @@
-from typing import Dict, Any
+from typing import Dict
 from hardware.base_controller import HardwareController
 
 class GestureActionMapper:
